Move RPC.split to pubsub.go file
